Break useTask into 4 different files
THis was done to place each hook on its own file, a file called useTask
should only exist if there's a hook called useTask to go with it,
otherwise it makes more sense to have everything broken apart.
